Tests/LibWeb/Text/expected/hit-test-border-radius.txt

Corner (5,5):
<DIV id="target">
Near corner (25,25):
<DIV id="target">
Center (100,100):
<DIV id="target">
